# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Temple TV JCTM Admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/history:
    get:
      operationId: getChatHistory
      tags: [chat]
      summary: Recent chat messages for a channel
      description: |
        Returns the most recent (non-deleted) chat messages for the given
        channel, in chronological order. Used ONLY for initial paint and for
        reconnect catch-up — live messages are pushed over the WebSocket
        gateway at `/api/chat/ws`. Polling this endpoint is forbidden by the
        chat contract.
      parameters:
        - in: query
          name: channelId
          required: false
          schema:
            type: string
            default: temple-tv-live
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatHistoryResponse"
...
components:
  schemas:
...
    ChatMessage:
      type: object
      description: A single chat message rendered to viewers.
      properties:
        id:
          type: string
          description: Server-assigned canonical id (uuid).
        channelId:
          type: string
        userId:
          type: string
          nullable: true
          description: Authenticated user id, or null for anonymous viewers.
        displayName:
          type: string
        body:
          type: string
        createdAtMs:
          type: integer
          format: int64
        broadcastItemId:
          type: string
          nullable: true
          description: Playback item that was airing when this message was sent (broadcast-aware history).
        broadcastItemTitle:
          type: string
          nullable: true
      required: [id, channelId, userId, displayName, body, createdAtMs, broadcastItemId, broadcastItemTitle]
    ChatHistoryResponse:
      type: object
      properties:
        channelId:
          type: string
        messages:
          type: array
          items: { $ref: "#/components/schemas/ChatMessage" }
      required: [channelId, messages]
...